Promote Shared Singleton Signer Implementation (#413)
This PR promotes the shared singleton signer implementation to the 4337
directory instead of being a "test" contract. This involves:
- Adding it to the deployments
- Moving and renaming the contract to remove the `Test` prefix

I also took this opportunity to rename it to a "shared" signer instead
of "singleton" signer, so that it does not get confused with the other
singleton contract (the `SafeWebAuthnSignerSingleton`) that already
exists in this repository.

Additionally, I moved some of the test files around in order to clearly
differentiate between E2E tests and unit tests in the `test/4337`
directory.

Unit tests for 100% coverage of the `SafeWebAuthnSharedSigner` will be
added in a follow up PR.
