+++ modules/passkey/src/deploy/passkey4337.ts
@@ -14,6 +14,12 @@ const deploy: DeployFunction = async ({ deployments, getNamedAccounts }) => {
     log: true,
     deterministicDeployment: true,
   })
+  await deploy('SafeWebAuthnSharedSigner', {
+    from: deployer,
+    args: [],
+    log: true,
+    deterministicDeployment: true,
+  })
 }
 
 deploy.dependencies = ['entrypoint']
